daemon: move KPR init logic to pkg/kpr/initializer package
This commit moves the KPR initialization logic from the daemon
to the package `pkg/kpr/initializer`.
