+++ pkg/kpr/initializer/kube_proxy_replacement_nodeport_linux_test.go
@@ -3,7 +3,7 @@
 
 //go:build linux
 
-package cmd
+package initializer
 
 import (
 	"fmt"
